chore: move ClientStub and InMemoryFeedStore to shared scope
