+++ EssentialApp/EssentialAppTests/FeedAcceptanceTests.swift
@@ -101,73 +101,3 @@ final class FeedAcceptanceTests: XCTestCase {
         ]])
     }
 }
-
-fileprivate class HTTPClientStub: HTTPClient {
-    private class Task: HTTPClientTask {
-        func cancel() { }
-    }
-    
-    private let stub: (URL) -> (HTTPClient.Result)
-    
-    init(stub: @escaping (URL) -> HTTPClient.Result) {
-        self.stub = stub
-    }
-    
-    func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> any EssentialFeed.HTTPClientTask {
-        completion(stub(url))
-        return Task()
-    }
-    
-    static var offline: HTTPClientStub {
-        HTTPClientStub { _ in .failure(NSError(domain: "offline", code: 0)) }
-    }
-    
-    static func online(_ stub: @escaping (URL) -> (Data, HTTPURLResponse)) -> HTTPClientStub {
-        HTTPClientStub { url in .success(stub(url)) }
-    }
-}
-
-fileprivate class InMemoryFeedStore: FeedStore, FeedImageDataStore {
-    private(set) var feedCache: CachedFeed?
-    private(set) var feedImageDataCache: [URL: Data] = [:]
-    
-    init(feedCache: CachedFeed? = nil) {
-        self.feedCache = feedCache
-    }
-    
-    static var empty: InMemoryFeedStore {
-        InMemoryFeedStore()
-    }
-    
-    static var withNonExpiredFeedCache: InMemoryFeedStore {
-        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date()))
-    }
-    
-    static var withExpiredFeedCache: InMemoryFeedStore {
-        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date.distantPast))
-    }
-    
-    func insert(_ data: Data, for url: URL, completion: @escaping (InsertionResult) -> Void) {
-        feedImageDataCache[url] = data
-        completion(.success(()))
-    }
-    
-    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
-        let data = feedImageDataCache[url]
-        completion(.success(data))
-    }
-    
-    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        feedCache = nil
-        completion(.success(()))
-    }
-    
-    func insert(_ feed: [EssentialFeed.LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        feedCache = CachedFeed(feed: feed, timestamp: timestamp)
-        completion(.success(()))
-    }
-    
-    func retrieve(completion: @escaping RetrievalCompletion) {
-        completion(.success(feedCache))
-    }
-}
