+++ EssentialApp/EssentialAppTests/Helpers/InMemoryFeedStore.swift
@@ -1,0 +1,54 @@
+//
+//  InMemoryFeedStore.swift
+//  EssentialAppTests
+//
+//  Created by Nikhil Menon on 6/18/24.
+//
+
+import Foundation
+import EssentialFeed
+
+class InMemoryFeedStore: FeedStore, FeedImageDataStore {
+    private(set) var feedCache: CachedFeed?
+    private(set) var feedImageDataCache: [URL: Data] = [:]
+    
+    init(feedCache: CachedFeed? = nil) {
+        self.feedCache = feedCache
+    }
+    
+    static var empty: InMemoryFeedStore {
+        InMemoryFeedStore()
+    }
+    
+    static var withNonExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date()))
+    }
+    
+    static var withExpiredFeedCache: InMemoryFeedStore {
+        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date.distantPast))
+    }
+    
+    func insert(_ data: Data, for url: URL, completion: @escaping (InsertionResult) -> Void) {
+        feedImageDataCache[url] = data
+        completion(.success(()))
+    }
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        let data = feedImageDataCache[url]
+        completion(.success(data))
+    }
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        feedCache = nil
+        completion(.success(()))
+    }
+    
+    func insert(_ feed: [EssentialFeed.LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        feedCache = CachedFeed(feed: feed, timestamp: timestamp)
+        completion(.success(()))
+    }
+    
+    func retrieve(completion: @escaping RetrievalCompletion) {
+        completion(.success(feedCache))
+    }
+}
